+++ integration-tests-scala/src/test/scala/tech/beshu/ror/integration/PluginTestSupport.scala
@@ -1,7 +1,7 @@
 package tech.beshu.ror.integration
 
-import tech.beshu.ror.utils.containers.generic.{CallingEsDirectly, EsClusterWithRorPluginProvider, TargetEsContainer}
+import tech.beshu.ror.utils.containers.generic.{CallingEsDirectly, EsContainerWithRorPluginCreator, TargetEsContainer}
 
-trait PluginTestSupport extends EsClusterWithRorPluginProvider with CallingEsDirectly {
+trait PluginTestSupport extends EsContainerWithRorPluginCreator with CallingEsDirectly {
   this: TargetEsContainer =>
 }
