+++ tests-utils/src/main/scala/tech/beshu/ror/utils/containers/generic/SingleContainerCreator.scala
@@ -1,0 +1,9 @@
+package tech.beshu.ror.utils.containers.generic
+
+import cats.data.NonEmptyList
+import monix.eval.Task
+
+trait SingleContainerCreator {
+
+  def create(name: String, nodeNames: NonEmptyList[String], clusterSettings: ClusterSettings): RorContainer
+}
